tryin to make close function, admin stopped working

+++ commerce/auctions/views.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import authenticate, login, logout
 from django.db import IntegrityError
 from django.http import HttpResponse, HttpResponseRedirect
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 from django.urls import reverse
 from django import forms
 
@@ -86,39 +86,51 @@ def create_listing(request):
 def listing_page(request, title):
     
     message = None
-    listing = Listing.objects.get(title=title)
-    
-    is_on_watchlist = Watchlist.objects.filter(listing=listing, user=request.user).exists()
-    
-    if is_on_watchlist:
-        message = "Remove from watchlist"
-    else:
-        message = "Add to watchlist"    
+    listing = get_object_or_404(Listing, title=title)
+    required_min_bid = listing.bid + 1
 
-    if request.method == "POST":
+    if request.user.is_authenticated:
+
+        is_on_watchlist = Watchlist.objects.filter(listing=listing, user=request.user).exists()
         
-        if "watchlist" in request.POST:
+        if is_on_watchlist:
+            message = "Remove from watchlist"
+        else:
+            message = "Add to watchlist"    
+
+        if request.method == "POST":
+            
+            if "watchlist" in request.POST:
+            
+                try:
+                    entry = Watchlist(listing=listing, user=request.user)
+                    entry.save()
+                    message = "Added to watchlist"
+            
+                except IntegrityError:
+                    pass
+            
+                if is_on_watchlist:
+                    Watchlist.objects.filter(listing=listing, user=request.user).delete()
+                    message = "Removed from watchlist"
+            
+            if "bid" in request.POST:
+                
+                if request.POST["bid_value"]:
+
+                    bid = float(request.POST["bid_value"])
+                
+                    if bid > listing.bid:
+                        listing.bid = bid
+                        listing.last_bidder = request.user
+                        listing.save()
         
-            try:
-                entry = Watchlist(listing=listing, user=request.user)
-                entry.save()
-                message = "Added to watchlist"
-        
-            except IntegrityError:
-                pass
-        
-            if is_on_watchlist:
-                Watchlist.objects.filter(listing=listing, user=request.user).delete()
-                message = "Removed from watchlist"
-        
-        if "bid" in request.POST:
-        
-            listing.bid = request.POST["bid_value"]
-            listing.last_bidder = request.user
-    
+        #if request.user == listing.user:
+            
+
     return render(request, "auctions/listing_page.html",{
         "listing": listing,
-        "message": message,
-        "current_bid": listing.bid
+        "watchlist_button_message": message,
+        "required_min_bid": required_min_bid
     })
